infer script is added and refactorring
