+++ configs/classification/cifar10_resnet18.py
@@ -101,13 +101,13 @@ config = Config(
     batch_size_train=100,
     batch_size_valid=200,
     lr=1e-1,
-    max_epoch=100,
+    max_epoch=2,
     train_transforms=train_basic_augs(crop_size=(32, 32)),
     valid_transforms=[],
     comment="cifar10_resnet18",
     train_loader_workers=8,
-    valid_loader_workers=10,
-    single_pass_length=1.0,
+    valid_loader_workers=8,
+    single_pass_length=0.01,
     resume_from_checkpoint=None,
     shuffle_train=True,
     output_config=[],
