+++ pylantern/classification/__main__.py
@@ -1,25 +1,12 @@
-import os
-import sys
-from enum import Enum
 from pathlib import Path
-from shutil import copy
-from typing import List, Optional
 
-from matches.accelerators import DDPAccelerator, VanillaAccelerator
-from matches.loop import Loop
-from matches.utils import unique_logdir
 import typer
 
-from ..config import load_config, dump_config_json, dump_config_txt
-from ..config_generator import ConfigGenerator, load_config_generator
-from ..common.utils import (
-    DevMode,
-    load_pickle,
-    dump_pickle,
-    prepare_logdir,
-    prepare_comment,
-    copy_config,
-    copy_config_generator,
+from ..common.utils import DevMode
+from ..common.train_utils import (
+    train_routine,
+    train_config_generator_routine,
+    infer_routine,
 )
 from .config import ClassificationConfig
 from .train_fns import train_fn, infer_fn
@@ -32,24 +19,18 @@ app = typer.Typer()
 def train(
     config_path: Path,
     comment: str = typer.Option(None, "--comment", "-C"),
-    logdir: Path = None,
+    logdir: Path = typer.Option(None, "--logdir", "-l"),
     gpus: str = typer.Option(None, "--gpus", "--gpu", "-g"),
     dev_mode: DevMode = typer.Option(DevMode.DISABLED, "--dev-mode", "-m"),
 ):
-    config: ClassificationConfig = load_config(config_path, ClassificationConfig)
-    comment, config = prepare_comment(comment, config_path, config)
-    logdir = prepare_logdir(logdir, comment)
-    copy_config(config_path, logdir)
-
-    loop = Loop(
-        logdir,
-        config.train_callbacks(dev_mode != dev_mode.DISABLED),
-        loader_override=dev_mode.value,
-    )
-    loop.launch(
-        train_fn,
-        DDPAccelerator(gpus) if gpus is not None else VanillaAccelerator("cpu"),
-        config=config,
+    train_routine(
+        config_path=config_path,
+        config_cls=ClassificationConfig,
+        train_fn=train_fn,
+        comment=comment,
+        logdir=logdir,
+        gpus=gpus,
+        dev_mode=dev_mode,
     )
 
 
@@ -61,57 +42,34 @@ def train_config_generator(
     gpus: str = typer.Option(None, "--gpus", "--gpu", "-g"),
     dev_mode: DevMode = typer.Option(DevMode.DISABLED, "--dev-mode", "-m"),
 ):
-    root_log_dir.mkdir(exist_ok=True, parents=True)
-    copy_config_generator(config_generator_path, root_log_dir)
-    config_generator: ConfigGenerator = load_config_generator(config_generator_path)
-    copy_config(config_generator.base_config_path, root_log_dir)
-
-    for idx, config in enumerate(config_generator()):
-        comment = f"{idx:02d}__{config.comment}"
-        if comment_postfix is not None:
-            comment += f"_{comment_postfix}"
-        config.comment = comment
-
-        logdir = root_log_dir / comment
-        logdir.mkdir(exist_ok=True, parents=True)
-        # dump_config_json(config, logdir / "config.json")
-        dump_config_txt(config, logdir / "config")
-        # dump_pickle(config, logdir / "config.pkl")
-
-        loop = Loop(
-            logdir,
-            config.train_callbacks(dev_mode != dev_mode.DISABLED),
-            loader_override=dev_mode.value,
-        )
-
-        loop.launch(
-            train_fn,
-            DDPAccelerator(gpus) if gpus is not None else VanillaAccelerator("cpu"),
-            config=config,
-        )
+    train_config_generator_routine(
+        config_generator_path=config_generator_path,
+        root_log_dir=root_log_dir,
+        train_fn=train_fn,
+        comment_postfix=comment_postfix,
+        gpus=gpus,
+        dev_mode=dev_mode,
+    )
 
 
 @app.command()
 def infer(
     config_path: Path,
-    logdir: Path,
-    checkpoint: str = typer.Option("best", "-c"),
+    logdir: Path = typer.Option(None, "--logdir", "-l"),
+    checkpoint: str = typer.Option("best", "--checkpoint", "-c"),
     gpus: str = typer.Option(None, "--gpus", "--gpu", "-g"),
-    data_root: Optional[Path] = None,
-    output_name: Optional[str] = None,
+    data_root: Path = typer.Option(None, "--data-root", "-d"),
+    output_name: str = typer.Option(None, "--output-name", "-o"),
 ):
-    config: ClassificationConfig = load_config(config_path, ClassificationConfig)
-    loop = Loop(
-        logdir,
-        config.valid_callbacks(),
-    )
-    loop.launch(
-        infer_fn,
-        DDPAccelerator(gpus) if gpus is not None else VanillaAccelerator("cpu"),
-        config=config,
+    infer_routine(
+        config_path=config_path,
+        config_cls=ClassificationConfig,
+        infer_fn=infer_fn,
+        logdir=logdir,
         checkpoint=checkpoint,
         data_root=data_root,
         output_name=output_name,
+        gpus=gpus,
     )
