+++ scripts/infer_classifier.sh
@@ -1,0 +1,1 @@
+python3 -m pylantern.classification infer logs/cifar10_resnet18__3/230106_1907/config.py -c "best" -o "infer_best"
